# Conflicts:
#	hedera-node/hedera-mono-service/src/main/java/com/hedera/node/app/service/mono/state/StateModule.java
#	hedera-node/hedera-mono-service/src/test/java/com/hedera/node/app/service/mono/ServicesAppTest.java
#	hedera-node/hedera-mono-service/src/test/java/com/hedera/node/app/service/mono/ServicesStateTest.java
#	platform-sdk/platform-apps/demos/CryptocurrencyDemo/src/main/java/com/swirlds/demo/crypto/CryptocurrencyDemoMain.java
#	platform-sdk/platform-apps/demos/HelloSwirldDemo/src/main/java/com/swirlds/demo/hello/HelloSwirldDemoMain.java
#	platform-sdk/platform-apps/demos/StatsDemo/src/main/java/com/swirlds/demo/stats/StatsDemoMain.java
#	platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/PlatformBuilder.java
#	platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/crypto/CryptoStatic.java
#	platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/crypto/EnhancedKeyStoreLoader.java
#	platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/SwirldsGui.java
#	platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/network/connectivity/OutboundConnectionCreator.java
#	platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/state/address/AddressBookNetworkUtils.java
#	platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/util/BootstrapUtils.java
#	platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/crypto/CryptoArgsProvider.java
#	platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/util/AddressBookNetworkUtilsTests.java

+++ hedera-node/hedera-config/src/main/java/com/hedera/node/config/data/ConsensusConfig.java
@@ -25,6 +25,6 @@ public record ConsensusConfig(
         @ConfigProperty(value = "message.maxBytesAllowed", defaultValue = "1024") @NetworkProperty
                 int messageMaxBytesAllowed,
         @ConfigProperty(value = "handle.maxPrecedingRecords", defaultValue = "3") @NetworkProperty
-                long handleMaxPrecedingRecords,
+                int handleMaxPrecedingRecords,
         @ConfigProperty(value = "handle.maxFollowingRecords", defaultValue = "50") @NetworkProperty
-                long handleMaxFollowingRecords) {}
+                int handleMaxFollowingRecords) {}
